Complete parse the init file

+++ collegeManager/Course.cpp
@@ -1,3 +1,3 @@
 #include "Course.h"
 
-Course::Course(long _id, std::string _name, int _creditPoints): id(_id), name(_name), creditPoints(_creditPoints) {}
+Course::Course(int _id, std::string _name, int _creditPoints): id(_id), name(_name), creditPoints(_creditPoints) {}
